Rename for consistency

+++ implementations/pytorch_conv/discretize_pde.py
@@ -10,7 +10,7 @@ def discretize_heat_equation(N:int, dt:float, t_end:float, write_every:int,devic
     os.makedirs("output", exist_ok=True)
     t = 0
     h = 1 / (N - 1)
-    u = torch.ones((N, N), device=device)
+    u = torch.ones((N, N), device=device, dtype=torch.float32)
     u[0:N, 0] = 0
     u[0:N, N - 1] = 0
     u[0, 0:N - 1] = 0
@@ -22,7 +22,7 @@ def discretize_heat_equation(N:int, dt:float, t_end:float, write_every:int,devic
         conv_layer = nn.Conv2d(in_channels=1, out_channels=1, kernel_size=3, bias=False)
 
         # Set custom values for weights
-        custom_weights = torch.tensor([[[[0., 1, 0], [1, -4, 1], [0, 1, 0]]]], device=device)
+        custom_weights = torch.tensor([[[[0., 1, 0], [1, -4, 1], [0, 1, 0]]]], device=device,dtype=torch.float32)
         conv_layer.weight = nn.Parameter(custom_weights)
         iteration=1
         for t in tqdm.tqdm(np.arange(0, t_end, dt)):
